Adjust warning fixes to be gcc-compatible
QVariant can't handle size_t on linux
_read etc don't actually exist under gcc despite claimed ISO standard
casting compatibility

+++ src/MEGASync/control/Preferences.h
@@ -436,7 +436,7 @@ public:
 
     static unsigned int PROXY_TEST_TIMEOUT_MS;
     static unsigned int MAX_IDLE_TIME_MS;
-    static size_t MAX_COMPLETED_ITEMS;
+    static unsigned int MAX_COMPLETED_ITEMS;
 
     static unsigned int MUTEX_STEALER_MS; //to create a task that steals the sdk mutex for a while (how long)
     static unsigned int MUTEX_STEALER_PERIOD_MS; //periodicity (how often)
