+++ src/MEGASync/gui/QCustomTransfersModel.cpp
@@ -103,7 +103,7 @@ void QCustomTransfersModel::onTransferFinish(MegaApi *api, MegaTransfer *t, Mega
                     TransferItemData *item = new TransferItemData(transfer);
                     item->data.publicNode = isPublicNode;
 
-                    if (transfers.size() == Preferences::MAX_COMPLETED_ITEMS)
+                    if ((int)transfers.size() == (int)Preferences::MAX_COMPLETED_ITEMS)
                     {
                         TransferItemData *itemData = transferOrder.back();
                         int row = int(transferOrder.size()) - 1;
